Refactor and tidy up checklist code
