--- components/InfoBox/HotspotDetails/Checklist/index.js
@@ -1,1 +1,0 @@
-export { default } from './Checklist'
